requirements.txt
pandas
streamlit
plotly
python-dotenv
matplotlib
koreanize-matplotlib
seaborn
scikit-learn
requests
